refactor: raplace main layout to widgets
