--- src/app/components/Footer.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Footer = () => {
-  return (
-    <div>
-
-    </div>
-  );
-};
-
-export default Footer;
