--- src/app/components/Header.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Header = () => {
-  return (
-    <header>
-      I am header
-    </header>
-  );
-};
-
-export default Header;
