+++ src/pages/_app.tsx
@@ -1,7 +1,7 @@
 import '@/shared/design/globals.scss';
 
 import type { AppProps } from 'next/app';
-import MainLayout from "@/app/components/MainLayout";
+import MainLayout from "@/widgets/layouts/main";
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
