+++ src/pages/index.tsx
@@ -1,3 +1,14 @@
+import React from "react";
+
 export default function Home() {
-  return <>My app</>;
+  return (
+    <>
+      <input className="graphql-input"/>
+      <div className="graphql-editor">
+        <div className="graphql-textarea-wrapper">
+          <textarea className="graphql-textarea"></textarea>
+        </div>
+        <div className="graphql-editor-right"></div>
+      </div>
+    </>);
 }
