+++ src/widgets/layouts/main/components/Footer/Footer.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Footer: React.FC = () => {
+  return <div></div>;
+};
