+++ src/widgets/layouts/main/components/Footer/index.tsx
@@ -1,0 +1,1 @@
+export * from './Footer';
