+++ src/widgets/layouts/main/components/Header/Header.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export const Header: React.FC = () => {
+  return <header>I am header</header>;
+};
